+++ resources/views/components/receipts.blade.php
@@ -1,16 +1,16 @@
 <x-jet-action-section>
+
     <x-slot name="title">
-        {{ __('Receipt Information') }}
+        {{ __('Receipts') }}
     </x-slot>
 
     <x-slot name="description">
-        {{ __('A list of transaction receipts.') }}
+        {{ __('A list of transactions and receipts.') }}
     </x-slot>
 
-    <x-slot name="content">
-    
+    <x-slot name="content">    
         <div class="text-gray-600">
-            <table width="100%">
+            <table width="100%" class="table-auto">
                 <thead class="bg-gray-50">
                     <tr>
                         <td nowrap><strong>ID</strong></th>
@@ -20,17 +20,19 @@
                         <td style="text-align:center"><strong>Billing Date</strong></th>
                     </tr>
                 </thead>
-                @foreach($receipts as $receipt)
-                <tr>
-                    <td>{{ $receipt->payfast_payment_id }}</td>
-                    <td nowrap>{{ $receipt->item_name }}</td>
-                    <td style="text-align:right">R {{ $receipt->amount_gross }}</td>
-                    <td style="padding-left:10px">{{ $receipt->payment_status }}</td>                    
-                    <td nowrap>{{ $receipt->billing_date->format('Y-m-d') }}</td>
-                </tr>
-                @endforeach
+                <tbody>
+                    @foreach($receipts as $receipt)                        
+                        <tr class="odd:bg-white even:bg-gray-50">
+                            <td>{{ $receipt->payfast_payment_id }}</td>
+                            <td>{{ $receipt->item_name }}</td>
+                            <td style="text-align:right">R {{ $receipt->amount_gross }}</td>
+                            <td style="padding-left:10px">{{ $receipt->payment_status }}</td>                    
+                            <td nowrap>{{ $receipt->billing_date->format('Y-m-d') }}</td>
+                        </tr>
+                    @endforeach
+                </tbody>
             </table>
-        </div>
-                                        
+        </div>                                        
     </x-slot>
+
 </x-jet-action-section>