+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use FintechSystems\PayFast\Components\Billing;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/payfast/return', function() {
@@ -11,4 +12,9 @@ Route::get('/payfast/cancel', function() {
 });
 
 Route::post('/payfast/notify', 'FintechSystems\PayFast\Http\Controllers\WebhookController');
+
 Route::post('/payfast/webhook', 'FintechSystems\PayFast\Http\Controllers\WebhookController');
+
+Route::middleware(['web', 'auth:sanctum', 'verified'])
+    ->get('/user/billing', Billing::class)
+    ->name('profile.billing');