+++ src/Components/Billing.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace FintechSystems\PayFast\Components;
+
+use Livewire\Component;
+
+class Billing extends Component
+{    
+    public function render()
+    {
+        return view('vendor.payfast.components.billing');
+    }
+}