+++ src/Components/Receipts.php
@@ -11,11 +11,11 @@ class Receipts extends Component
 
     public $receipts;
 
+    protected $listeners = ['refreshComponent' => '$refresh'];
+
     public function mount()
     {
-        $this->user = Auth::user();
-
-        $this->receipts = $this->user->receipts;
+        $this->user = Auth::user();        
     }
 
     /**
@@ -25,6 +25,8 @@ class Receipts extends Component
      */
     public function render()
     {
+        $this->receipts = $this->user->receipts;
+        
         return view('vendor.payfast.components.receipts');
     }
 }