+++ src/Components/Subscriptions.php
@@ -23,14 +23,15 @@ class Subscriptions extends Component
 
     public $mergeFields;
 
-    protected $listeners = [
-        // 'billingUpdated' => '$refresh',
+    protected $listeners = [        
         'billingUpdated' => 'billingWasUpdated',
     ];
 
     public function billingWasUpdated()
     {
-        ray("billingWasUpdated event fired, resetting displayingCreateSubscription");
+        ray("billingUpdated event fired, refreshing receipts and removing subscription display");
+
+        $this->emitTo('receipts', 'refreshComponent');
 
         $this->displayingCreateSubscription = false;
     }
@@ -48,8 +49,7 @@ class Subscriptions extends Component
 
     public function cancelSubscription()
     {
-        Log::info('Cancelling subscription for ' . $this->user->subscriptions()->active()->first()->payfast_token);
-        // ray($this->user->subscriptions()->active()->first()->payfast_token);
+        ray('Cancelling subscription for ' . $this->user->subscriptions()->active()->first()->payfast_token)->orange();        
 
         $this->user->subscription('default')->cancel2();
 
@@ -65,9 +65,10 @@ class Subscriptions extends Component
     {
         $payfast_token = $this->user->subscription('default')->payfast_token;
 
-        ray("The updateCard token is ", $payfast_token);
-
-        $url = "https://www.payfast.co.za/eng/recurring/update/$payfast_token?return=" . config('app.url') . "/user/profile?card_updated=true";
+        ray("updateCard has been called with this token: $payfast_token");
+        
+        // $url = "https://www.payfast.co.za/eng/recurring/update/$payfast_token?return=" . config('app.url') . "/user/profile?card_updated=true";
+        $url = PayFast::url() . "/recurring/update/$payfast_token?return=" . PayFast::callbackUrl() . "/user/profile?card_updated=true";
 
         $message = "updateCard is going to redirect()->to this URL: " . $url;
 
@@ -92,7 +93,7 @@ class Subscriptions extends Component
 
     public function displayCreateSubscription()
     {
-        if ($this->user->onGenericTrial()) {
+        if ($this->user->onGenericTrial()) {            
             $subscriptionStartsAt = $this->user->trialEndsAt()->addDay()->format('Y-m-d');
 
             $this->mergeFields = array_merge($this->mergeFields, ['amount' => 0]);