+++ src/PayFastServiceProvider.php
@@ -2,10 +2,12 @@
 
 namespace FintechSystems\PayFast;
 
+use Livewire\Livewire;
+use Illuminate\Support\ServiceProvider;
+use FintechSystems\PayFast\Components\Banner;
+use FintechSystems\PayFast\Components\Billing;
 use FintechSystems\PayFast\Components\Receipts;
 use FintechSystems\PayFast\Components\Subscriptions;
-use Illuminate\Support\ServiceProvider;
-use Livewire\Livewire;
 
 class PayFastServiceProvider extends ServiceProvider
 {
@@ -30,6 +32,10 @@ class PayFastServiceProvider extends ServiceProvider
         Livewire::component('subscriptions', Subscriptions::class);
 
         Livewire::component('receipts', Receipts::class);
+
+        Livewire::component('banner', Banner::class);
+
+        Livewire::component('billing', Billing::class);
     }
 
     public function register()