+++ tests/Feature/WebhooksTest.php
@@ -16,7 +16,7 @@ class WebhooksTest extends FeatureTestCase
     public function test_gracefully_handle_webhook_without_alert_name()
     {
         $this->postJson('payfast/webhook', [
-            'event_time' => now()->addDay()->format('Y-m-d H:i:s'),
+            'ping' => now(),
         ])->assertOk();
     }